+++ src/app/models/text-analysis-input.model.ts
@@ -1,0 +1,7 @@
+export interface TextAnalysisInput {
+  readonly text: string;
+}
+
+export interface NormalizedTextAnalysisInput extends TextAnalysisInput {
+  readonly normalizedText: string;
+}